Some other performance optimizations

+++ ic2-reactors/src/main/groovy/com/gmail/dario/reactors/components/reflectors/NeutronReflector.groovy
@@ -1,5 +1,5 @@
 package com.gmail.dario.reactors.components.reflectors
 
 class NeutronReflector extends Reflector {
-    final BigDecimal maxPulses = 20_000
+    final int maxPulses = 20_000
 }
