+++ ic2-reactors/src/main/groovy/com/gmail/dario/reactors/components/reflectors/ThickNeutronReflector.groovy
@@ -1,5 +1,5 @@
 package com.gmail.dario.reactors.components.reflectors
 
 class ThickNeutronReflector extends Reflector {
-    final BigDecimal maxPulses = 120_000
+    final int maxPulses = 120_000
 }
